Update sqlc.yaml and helpers.go, delete db/querier.go, and modify docker-compose.yml

+++ cmd/api/helpers.go
@@ -38,13 +38,13 @@ func (app *application) writeJSON(w http.ResponseWriter, status int, data envelo
 	return nil
 }
 
-func (app *application) readIDParam(r *http.Request) (int64, error) {
+func (app *application) readIDParam(r *http.Request) (int32, error) {
 	idStr := chi.URLParam(r, "id")
 	id, err := strconv.ParseInt(idStr, 10, 64)
 	if err != nil || id < 1 {
 		return 0, errors.New("invalid id parameter")
 	}
-	return id, nil
+	return int32(id), nil
 }
 
 func (app *application) readJSON(w http.ResponseWriter, r *http.Request, dst any) error {
